test: 인수테스트 추가

+++ backend/src/main/java/com/woowacourse/gongseek/article/domain/repository/ArticleRepositoryCustom.java
@@ -1,7 +1,6 @@
 package com.woowacourse.gongseek.article.domain.repository;
 
 import com.woowacourse.gongseek.article.domain.Article;
-import com.woowacourse.gongseek.article.presentation.dto.ArticlePageResponse;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
@@ -10,7 +9,7 @@ public interface ArticleRepositoryCustom {
 
     List<Article> findAllByPage(Long cursorId, Integer views, String category, String sortType, int pageSize);
 
-    Slice<Article> findAllByLikes(Long cursorId, Integer likes, String category, Pageable pageable);
+    Slice<Article> findAllByLikes(Long cursorId, Long likes, String category, Pageable pageable);
 
     List<Article> searchByContainingText(Long cursorId, int pageSize, String searchText);
